Move tests out to a file
This is slightly better, as when you modify the tests, Cargo is smart enough to know that the library isn't modified, and so won't try and recompile it.
